pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simpleorm"
version = "0.1.0"
description = "A simple ORM library"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2",
    "pandas",
    "psycopg2-binary",
]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["orm", "database", "sql"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

[project.urls]
Homepage = "https://github.com/yourusername/simpleorm"
Documentation = "https://github.com/yourusername/simpleorm#readme"
Repository = "https://github.com/yourusername/simpleorm"

[tool.setuptools.packages.find]
where = ["."]
include = ["simpleorm*"]

[project.optional-dependencies]
dev = [
    "build>=0.10",
    "twine>=4.0",
]
